adecuaciones en migraciones para el nuevo tipo de dato usado en laravel > 6 en campos autoincrementales (claves primarias)

+++ database/migrations/2015_01_26_115523_create_permission_role_table.php
@@ -14,10 +14,10 @@ class CreatePermissionRoleTable extends Migration
     {
         if (!Schema::hasTable('permission_role')) {
             Schema::create('permission_role', function (Blueprint $table) {
-                $table->increments('id')->unsigned();
-                $table->integer('permission_id')->unsigned()->index();
+                $table->bigIncrements('id')->unsigned();
+                $table->bigInteger('permission_id')->unsigned()->index();
                 $table->foreign('permission_id')->references('id')->on('permissions')->onDelete('cascade');
-                $table->integer('role_id')->unsigned()->index();
+                $table->bigInteger('role_id')->unsigned()->index();
                 $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
                 $table->timestamps();
             });
