+++ database/migrations/2018_03_31_155715_create_parishes_table.php
@@ -15,10 +15,10 @@ class CreateParishesTable extends Migration
     {
         if (!Schema::hasTable('parishes')) {
             Schema::create('parishes', function (Blueprint $table) {
-                $table->increments('id')->comment('Identificador único del registro');
+                $table->bigIncrements('id')->comment('Identificador único del registro');
                 $table->string('name', 100)->comment('Nombre de la Parroquia');
                 $table->string('code', 10)->nullable()->comment('Código que identifica a la Parroquia');
-                $table->integer('municipality_id')->unsigned()
+                $table->bigInteger('municipality_id')->unsigned()
                       ->comment('Identificador asociado al Municipio al que pertenece');
                 $table->foreign('municipality_id')->references('id')->on('municipalities')
                       ->onDelete('restrict')->onUpdate('cascade');
