+++ database/migrations/2019_09_21_162701_create_deductions_table.php
@@ -20,7 +20,7 @@ class CreateDeductionsTable extends Migration
                 $table->text('description')->nullable()->comment("Descripción de la deducción");
                 $table->text('formula')->comment("Fórmula de la deducción a aplicar");
                 $table->boolean('active')->default(true)->comment('Establece si la deducción esta activa o no');
-                $table->integer('accounting_account_id')->nullable()->unsigned()
+                $table->bigInteger('accounting_account_id')->nullable()->unsigned()
                       ->comment('Identificador asociado a la cuenta contable');
                 $table->foreign('accounting_account_id')->references('id')
                       ->on('accounting_accounts')->onUpdate('restrict')->onDelete('restrict');
