+++ modules/Asset/Database/Migrations/2019_09_02_164151_create_assets_table.php
@@ -27,46 +27,46 @@ class CreateAssetsTable extends Migration
     {
         if (!Schema::hasTable('assets')) {
             Schema::create('assets', function (Blueprint $table) {
-                $table->increments('id')->comment('Identificador único del registro');
+                $table->bigIncrements('id')->comment('Identificador único del registro');
 
-                $table->integer('asset_type_id')->unsigned()
+                $table->bigInteger('asset_type_id')->unsigned()
                       ->comment('Identificador único del tipo de bien (1 Mueble, 2 Inmueble)');
                 $table->foreign('asset_type_id')->references('id')->on('asset_types')
                       ->onDelete('restrict')->onUpdate('cascade');
-                
-                $table->integer('asset_category_id')->unsigned()
+
+                $table->bigInteger('asset_category_id')->unsigned()
                       ->comment('Identificador único de la categoria general a la que pertence el bien');
                 $table->foreign('asset_category_id')->references('id')->on('asset_categories')
                       ->onDelete('restrict')->onUpdate('cascade');
 
-                $table->integer('asset_subcategory_id')->unsigned()
+                $table->bigInteger('asset_subcategory_id')->unsigned()
                       ->comment('Identificador único de la subcategoria a la que pertence el bien');
                 $table->foreign('asset_subcategory_id')->references('id')->on('asset_subcategories')
                       ->onDelete('restrict')->onUpdate('cascade');
 
-                $table->integer('asset_specific_category_id')->unsigned()
+                $table->bigInteger('asset_specific_category_id')->unsigned()
                       ->comment('Identificador único de la categoria especifica a la que pertence el bien');
                 $table->foreign('asset_specific_category_id')->references('id')->on('asset_specific_categories')
                       ->onDelete('restrict')->onUpdate('cascade');
 
-                $table->integer('asset_condition_id')->nullable()->unsigned()
+                $table->bigInteger('asset_condition_id')->nullable()->unsigned()
                       ->comment('Identificador único de la condicion física del bien');
                 $table->foreign('asset_condition_id')->references('id')->on('asset_conditions')
                       ->onDelete('restrict')->onUpdate('cascade');
 
-                $table->integer('asset_acquisition_type_id')->nullable()->unsigned()
+                $table->bigInteger('asset_acquisition_type_id')->nullable()->unsigned()
                       ->comment('Identificador único del tipo de adquisicion del bien');
                 $table->foreign('asset_acquisition_type_id')->references('id')->on('asset_acquisition_types')
                       ->onDelete('restrict')->onUpdate('cascade');
-                
+
                 $table->year('acquisition_year')->nullable()->unsigned()->comment('Año de adquisicion del bien');
-                
-                $table->integer('asset_status_id')->nullable()->unsigned()
+
+                $table->bigInteger('asset_status_id')->nullable()->unsigned()
                       ->comment('Identificador único del estatus de uso del bien');
                 $table->foreign('asset_status_id')->references('id')->on('asset_status')
                       ->onDelete('restrict')->onUpdate('cascade');
 
-                $table->integer('asset_use_function_id')->nullable()->unsigned()
+                $table->bigInteger('asset_use_function_id')->nullable()->unsigned()
                       ->comment('Identificador único de la función de uso del bien (solo para bienes inmuebles)');
                 $table->foreign('asset_use_function_id')->references('id')->on('asset_use_functions')
                       ->onDelete('restrict')->onUpdate('cascade');
@@ -82,19 +82,19 @@ class CreateAssetsTable extends Migration
 
                 $table->float('value')->nullable()->unsigned()
                       ->comment('Valor en libros del bien');
-                $table->integer('currency_id')->unsigned()->nullable()
+                $table->bigInteger('currency_id')->unsigned()->nullable()
                       ->comment('Identificador único asociado a la moneda');
                 $table->foreign('currency_id')->references('id')->on('currencies')
                       ->onDelete('restrict')->onUpdate('cascade');
 
                 $table->text('specifications')->nullable()->comment('Especificaciones técnicas del bien');
-                
+
                 $table->text('address')->nullable()->comment('Dirección fisíca de bien');
-                $table->integer('parish_id')->nullable()->unsigned()
+                $table->bigInteger('parish_id')->nullable()->unsigned()
                       ->comment('Identificador único de la parroquia donde esta ubicado el bien');
                 $table->foreign('parish_id')->references('id')->on('parishes')
                       ->onDelete('restrict')->onUpdate('cascade');
-                
+
                 $table->timestamps();
                 $table->softDeletes()->comment('Fecha y hora en la que el registro fue eliminado');
             });
