+++ modules/Asset/Database/Migrations/2019_09_02_164642_create_asset_request_deliveries_table.php
@@ -27,17 +27,17 @@ class CreateAssetRequestDeliveriesTable extends Migration
     {
         if (!Schema::hasTable('asset_request_deliveries')) {
             Schema::create('asset_request_deliveries', function (Blueprint $table) {
-                $table->increments('id')->comment('Identificador único del registro');
-                
+                $table->bigIncrements('id')->comment('Identificador único del registro');
+
                 $table->string('state')->nullable()->comment('Estado de la solicitud');
                 $table->text('observation')->nullable()->comment('Observaciones de la entrega');
-                
-                $table->integer('asset_request_id')
+
+                $table->bigInteger('asset_request_id')
                       ->comment('Identificador único de la solicitud asociada a la entrega');
                 $table->foreign('asset_request_id')->references('id')->on('asset_requests')
                       ->onDelete('restrict')->onUpdate('cascade');
-                
-                $table->integer('user_id')->comment('Identificador único del usuario que solicita la entrega');
+
+                $table->bigInteger('user_id')->comment('Identificador único del usuario que solicita la entrega');
                 $table->foreign('user_id')->references('id')->on('users')->onDelete('restrict')->onUpdate('cascade');
 
                 $table->timestamps();
